support empty compression stub
both comp-lzo and compress options are deprecated for the 2.5.x series.

from the manpage:

note: the stub (or empty) option is NOT compatible with the older option --comp-lzo no.

both disable compression, but the "compress" option is 0xfb, sending the
first byte of the payload to the last position of the buffer. note that
this is *after* padding.

moving forward, I'll have to implement more options, and look into
moving to the V2 framing. no intention to implement lzo or lz4 though.