let's embrace gvisor netstack
this is still a work in progress, but I think there's no way back at
this point.

without bringing a userspace tcpip stack, we'll be limited to
handcrafted icmp packets and the likes (or to require net caps to be
able to use raw sockets).

I'm slightly concerned about importing the whole gVisor monster, but for
what I read the lazy loading in go 1.17 was the reason why the separate
module was deprecated, so everything should be fine.

kudos to jason donenfeld for showing how elegant the coupling of the
userspace tun device with protocol code can be!