refactor: newMuxer & data channel
in the end I think the implementation managed to be simpler, in the
sense of not requiring a goroutine for handling the incoming loop.

this design has one severe limitation: it needs to be driven by user
reads. this can perhaps be solved by providing a Ticker that schedules
reads every given interval (so that we can make sure to respond to pings
and the server does not close our session).

other abstractions have been incorporated into the design (separation of
mutable state and behavior).

this commit also simplifies the TLS transport