simplify dialer
After review today with bassosimone, I think much of the comlexity of
the dialer can be removed. I suspect some of references that were kept
in the dialers do not belong there: they probably were introduced while
exploring the socks5 proxy (which I am not including in the tests right
now).

if the application wants to reuse the client, it's free to call
RawDialer.dial(), and keep the conn open, but it's not the
responsibility of the dialer struct to keep a reference to
the client itself.