bug: index out of range in parseServerHardResetPacket
MIV-01-001: Possible DoS via index out of range (Low)
During the fuzzing process of the minivpn/vpn package, it was found that
the parseServerHardResetPacket function fails to access the first
element of the serverHardReset payload when the payload is empty. This
led to the following crash error: “panic: runtime error: index out of
range [0] with length 0”. Please note that the
parseServerHardResetPacket function is used in association with the
newServerHardReset method which checks that the payload is not empty and
returns an error otherwise.

Thanks to 7asecurity that uncovered this bug during their Red Lab
engagement.

- Reference: MIV-01-001