bug: integer division by zero
During the fuzzing process of the minivpn/vpn package, it was found that
the bytesPadPKCS7 function fails to perform a modulo operation when the
blockSize argument is zero. This led to the following crash error:
“panic: runtime error: integer divide by zero”. This issue affects
projects using minivpn/vpn as a third party library or copying the
vulnerable code into another project, if an attacker is able to control
the blockSize value.

Thanks to 7asecurity that uncovered this bug during their Red
Lab engagement.

- Reference: MIV-01-002