bug: index out of range in maybeAddCompressPadding
MIV-01-004: Possible DoS via Index Out of Range (Medium)
During the fuzzing process of the minivpn/vpn package, it was found that the
maybeAddCompressPadding function fails to validate access to the array of byte
location provided as an argument. This led to the following crash error:
“panic: runtime error: index out of range [-1]”. Please note the
EncryptAndEncodePayload function, which invokes maybeAddCompressPadding, fails
to perform the length check as well.

Thanks to 7asecurity that uncovered this bug during their Red Lab engagement.

- Reference: MIV-01-004