bug: incorrect boundary check for non-aead ciphers
MIV-01-005: Possible DoS via Slice Bounds Out of Range (High)
During the fuzzing process of the minivpn/vpn package, it was found that the
decodeEncryptedPayloadNonAEAD function fails to perform slicing when buf is not
long enough. This led to the following crash error: “panic: runtime error:
slice bounds out of range [:36] with capacity 32”.

Thanks to 7asecurity that uncovered this bug during their Red
Lab engagement.

- Reference: MIV-01-005