feat: streamline support of obfs4 dialer
Before, there was a separate entry point for the obfs4 dialer. This
commit will create the right dialer if a line with proxy-obfs4 is added
to the openvpn config file (or the equivalent fields are set in
Options).

We also allow to pass any arbitrary dialer to the obfs4 base dialer.

While doing this, some attention is given to the thought that we might
want to support an arbitrary number of similar obfuscation proxies
-that's why we're trying to reuse the abstractions from the gost
project.