feat: try a range of ports if the required port is in use
Here I add a simple retry strategy that increments the port by one if
the default or configured listening port is in use.

While used as a library, it is be the caller's responsibility to
catch the error and retry accordingly, this retry strategy is only
provided for convenience while using the reference executable.

- Reference: MIV-01-007