feat: use P_DATA_V2 format
As pointed out by the security audit, the use of P_DATA_V1 format was
too conspicuous. In order to be able to use the v2 format, we needed to
implement a few requirements:

- client needs to push a specific IV_PROTO bit, so that the server knows
  that they have to assign us a peer-id.
- client needs to parse the pushed peer-id from the options pushed by
  the server.
- this peer-id is needed to encrypt (is passed to the authenticated data
  in GCM/AEAD mode) and to decrypt data payloads.

As a note, the peer-id expires in the server some time after the client
has terminated the session.

Useful pointers:

https://build.openvpn.net/doxygen/group__data__crypto.html
https://sourceforge.net/p/openvpn/mailman/message/32979583/
https://community.openvpn.net/openvpn/attachment/ticket/268/0001-Always-push-basic-set-of-peer-info-values-to-server-v2.patch