feat: add ability to trace handshake
In order to do that, we refactor the configuration handling. We move
options to openvpn options, and we create a config object that is
broader in scope.

This commit also updates the binary entrypoint to test this
functionality.

Since I was here, I added the pinger too to exercise the pinger over the
tunnel.